Statistics loader implementation
Added a json utils class to load json files and convert object to json and vice versa.
Added the car status json and the driver status json.
Fixed a typo in the driver status class and added a new attribute to the class.
Added a File constants class to store the json files path.
Integrate the data load in the car statistic handler.